requirements.txt
annotated-types==0.7.0
anyio==4.4.0
atproto==0.0.52
autopep8==2.3.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
cssselect2==0.7.0
defusedxml==0.7.1
dnspython==2.6.1
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
idna==3.8
Levenshtein==0.25.1
libipld==1.2.3
pillow==10.4.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-Levenshtein==0.25.1
rapidfuzz==3.9.7
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
tinycss2==1.3.0
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
websockets==13.0.1
